refactor: restore Normalize function name
