+++ AGENTS.md
@@ -22,7 +22,7 @@ make vet         # Run go vet
 
 # Run tests
 go test -v
-go test -v -run TestNormalizeFilename  # Run specific test
+go test -v -run TestNormalize  # Run specific test
 
 # Build for all platforms
 make build-all
@@ -45,8 +45,8 @@ go run . [flags] file1 [file2 ...]
 This project is organized into separate command and library packages:
 
 - **cmd/fnorm/main.go**: CLI entry point with flag parsing and file processing logic
-- **pkg/fnorm/normalize.go**: Library package exporting the `Normalize` function
-- **pkg/fnorm/normalize_test.go**: Table-driven tests for the normalization logic
+- **internal/normalize/normalize.go**: Library package exporting the `Normalize` function
+- **internal/normalize/normalize_test.go**: Table-driven tests for the normalization logic
 - **tools.go**: Development tool dependencies (build tag: tools)
 - **.golangci.yml**: Linter configuration
