+++ internal/normalize/example_test.go
@@ -1,0 +1,13 @@
+package normalize_test
+
+import (
+	"fmt"
+
+	normalize "github.com/laydros/fnorm/internal/normalize" //nolint:depguard // internal package import
+)
+
+func ExampleNormalize() {
+	result := normalize.Normalize("My File.PDF")
+	fmt.Println(result)
+	// Output: my-file.pdf
+}
