+++ internal/normalize/normalize_bench_test.go
@@ -1,8 +1,8 @@
-package fnorm
+package normalize
 
 import "testing"
 
-func BenchmarkNormalizeFilename(b *testing.B) {
+func BenchmarkNormalize(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		Normalize("My Complex File Name (Copy) #1.PDF")
 	}
