+++ internal/normalize/normalize_test.go
@@ -1,4 +1,4 @@
-package fnorm
+package normalize
 
 import "testing"
